+++ RELEASE_NOTES.md
@@ -20,6 +20,7 @@ This version moves service implementations into their related namespaces, to mak
 * `Keyboard.KeyboardServices` has been renamed to `Keyboard.Services`.
 * `Keyboard.KeyboardState` has been renamed to `Keyboard.State`.
 * `StandardKeyboardBehavior` has been renamed to `Keyboard.StandardBehavior`.
+* `StandardKeyboardActionHandler` has been renamed to `KeyboardAction.StandardHandler`.